refactor(swarm-accounting): preserve typed accounting error discriminants
Three settlement error-model cleanups so failure discriminants survive
the metric boundary instead of collapsing into opaque strings.

Remove the dead pseudosettle wire timestamp validator. The protocol read
path never validates ack timestamps, so the test-only validate_timestamp
helper and the unused PseudosettleError::InvalidTimestamp variant were dead;
drop both and assert ack recency directly in the retained test.

Move AccountingError into vertex-swarm-api and add a typed
SwarmError::AccountingDecision(#[from] AccountingError) variant. The
prepare_receive/prepare_provide boundary previously mapped every typed
accounting failure into the generic Accounting variant, flattening the
discriminant into a single accounting metric label; the typed variant now
carries it so the inner accounting_errors_total reason label survives. The
type lives in the api crate because accounting/core depends on api, not the
reverse, so a #[from] in api cannot reach down into accounting/core;
accounting/core re-exports it for the original path. The generic Accounting
variant and its constructors are kept additively.

The String escape-hatch variants (NetworkError, SigningFailed,
ValidationFailed, ChequeError::SignatureRecovery, AccountingError::
SettlementFailed) are left opaque: each either wraps a non-Clone underlying
error while the enum derives Clone, wraps multiple heterogeneous source
types plus a sourceless case, or carries no source type at all. Every one
already emits a clean snake_case strum label.
