+++ crates/swarm/api/src/lib.rs
@@ -55,7 +55,8 @@ pub use self::config::{
     SwarmStorerLaunchConfig, estimate_chunks_for_bytes, estimate_storage_bytes,
 };
 pub use self::error::{
-    ConfigAddressKind, ConfigError, ConfigResult, IdentityError, SwarmError, SwarmResult,
+    AccountingError, ConfigAddressKind, ConfigError, ConfigResult, IdentityError, SwarmError,
+    SwarmResult,
 };
 pub use self::identity::SwarmIdentity;
 pub use self::protocol::SwarmProtocol;
